+++ .github/workflows/pull-request.yml
@@ -1,9 +1,12 @@
 # This workflow will build a Java project with Maven
 # For more information see: https://help.github.com/actions/language-and-framework-guides/building-and-testing-java-with-maven
 
-name: Build and Test Extention
+name: Build and Quick Test Pull Request
 
-on: [ pull_request]
+on:
+  pull_request:
+    branches-ignore:
+      - 'dependabot/maven/org.liquibase-liquibase-core-*'
 
 jobs:
   test:
